Fix: Ajustes no problema com merge do repositorio/fork

+++ src/models/credit-request.model.ts
@@ -15,7 +15,7 @@ export class CreditRequestModel {
     @JoinColumn({ name: 'companyId', referencedColumnName: 'companyId' })
     company?: CompanyModel
 
-    @Column('float', { precision: 8, scale: 2 })
+    @Column('float', { precision: 16, scale: 2 })
     value: number
 
     @Column('varchar', { length: 30 })
